DBZ-9004: Remove ConnectionValidationUtil
